Organise the classes into suitable java packages
